feat: add car, auth

+++ lib/data/models/car.dart
@@ -1,8 +1,8 @@
 class Car {
   final String model;
-  final double distance;
-  final double fuelCapacity;
-  final double pricePerHour;
+  final int distance;
+  final int fuelCapacity;
+  final int pricePerHour;
 
   Car({required this.model, required this.distance, required this.fuelCapacity, required this.pricePerHour});
